refactor header for better RSC composition
